Minor "user-page" code cleanup
